+++ src/javascript/0070-Parser_.js
@@ -27,26 +27,25 @@ s.loadVars = function(time) {
     return;
 }
 
-/*
-A Parser object manages the conversion of the given passage text to the
-actual passage text used by SugarCube to display information.
-
-There is one instance (`st.parser`, see Page.js), built on
-State.variables, so its state can be stored in SugarCube's history.
-(This is needed, because the parser tracks dynamically generated text
-substitutions.) The class must therefore be made compatible with
-SugarCube by having `clone()` and `toJSON()` methods, and no recursive
-objects or object sharing. To achieve this, we also require it to have a
-constructor with no arguments and to have all its properties be SC
-supported types.
-
-There are various methods for processing different kinds of markup. The
-method, `procMarkup()`, runs them all in the appropriate order. Through
-`Config.passages.onProcess`, we instruct SugarCube to apply
-`procMarkup()` to passage text whenever a passage is rendered.
-*/
-
 s.Parser = function() {
+    /*
+    A Parser object manages the conversion of the given passage text to
+    the actual passage text used by SugarCube to display information.
+
+    There is one instance (`st.parser`, see Page.js), built on
+    State.variables, so its state can be stored in SugarCube's history.
+    (This is needed, because the parser tracks dynamically generated
+    text substitutions.) The class must therefore be made compatible
+    with SugarCube by having `clone()` and `toJSON()` methods, and no
+    recursive objects or object sharing. To achieve this, we also
+    require it to have a constructor with no arguments and to have all
+    its properties be SC supported types.
+
+    There are various methods for processing different kinds of markup.
+    The method, `procMarkup()`, runs them all in the appropriate order.
+    Through `Config.passages.onProcess`, we instruct SugarCube to apply
+    `procMarkup()` to passage text whenever a passage is rendered.
+    */
     this._textSubMap = new Map(); // Maps passage titles to arrays of
                                   // text subs. A text sub can be a
                                   // simple string or a functions